backend/requirements.txt
psycopg2-binary
gunicorn==21.2.0
djangorestframework-simplejwt